refactor: split market data slice logic into separate files
